fixed according to comments on previous pr
