+++ package.json
@@ -17,7 +17,7 @@
     "my-cli-tool": "./dist/esbuild/cli.js"
   },
   "scripts": {
-    "dev": "ts-node src/app.ts",
+    "dev": "ts-node src/index.ts",
     "lint": "eslint src/ --ext .js,.jsx,.ts,.tsx",
     "test": "jest",
     "clean": "rm -rf dist build package",
@@ -34,6 +34,8 @@
     "@types/node": "^16.3.2",
     "@typescript-eslint/eslint-plugin": "^4.19.0",
     "@typescript-eslint/parser": "^4.19.0",
+    "@types/express": "^4.17.13",
+    "@types/validator": "^13.6.3",
     "esbuild": "^0.11.11",
     "eslint": "^7.22.0",
     "jest": "^26.6.3",
@@ -43,8 +45,6 @@
     "typescript": "^4.2.3"
   },
   "dependencies": {
-    "@types/express": "^4.17.13",
-    "@types/validator": "^13.6.3",
     "express": "^4.17.1",
     "mysql2": "^2.2.5",
     "sequelize": "^6.6.5"
