+++ src/index.ts
@@ -1,16 +1,9 @@
 #!/usr/bin/env node
-import express from 'express';
 
 import database from './database/index'
-import mainRouter from './routes/index';
+import app from './server/app'
 
-const app = express();
 const port = 8080;
-
-const loggerMiddleware = (request: express.Request, _response: express.Response, next: express.NextFunction) => {
-  console.log(`${request.method} ${request.path}`);
-  next();
-}
 
 const authenticateDatabase = async () => {
   try {
@@ -23,16 +16,10 @@ const authenticateDatabase = async () => {
 }
 
 const init = () => {
-
   authenticateDatabase();
-  app.use(loggerMiddleware);
-
   app.listen(port, () => {
     console.log(`Hello World! Running on port ${port}`);
-  });
-  
-  app.use('/', mainRouter);
-  
+  }); 
 }
 
 init();
